fix/disable warnings. there are too many W4 and "potentially uninitialized", so those are disabled by 0ad_warning_disable.h.
the silly "int x = strlen" and very dangerous "int x = (void*)p" (and vice versa) problems are fixed.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5526 3db68df2-c116-0410-a063-a993310a9797

+++ source/dcdt/se/sr_array_pt.cpp
@@ -1,4 +1,5 @@
 #include "precompiled.h"
+#include "0ad_warning_disable.h"
 # include "sr_array_pt.h"
 
 //====================== SrArrayPtBase ==========================
@@ -134,7 +135,7 @@ static SrClassManagerBase* StaticManager = 0; // This is not thread safe...
 
 static int fcmp ( const void* pt1, const void* pt2 )
  {
-   typedef const int* cint;
+   typedef const intptr_t* cint;
    return StaticManager->compare( (const void*)*cint(pt1), (const void*)*cint(pt2) );
  }
